backend/requirements.txt
# FastAPI & Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart  # For file uploads

# Monitoring & Observability
mlflow[extras]

# Database & Storage
boto3
psycopg2-binary  # PostgreSQL adapter
asyncpg  # Async PostgreSQL adapter for chat persistence
redis>=5.0.0     # Redis client with async support
redis[hiredis]   # High-performance Redis parser
chromadb>=0.4.0  # Vector database
neo4j>=5.0.0     # Graph database driver

# Agent & AI Framework
langchain>=0.1.0
langchain-openai>=0.0.5  # ChatOpenAI for LangChain v1 alpha
langchain-anthropic>=0.1.0  # Claude fallback option
langchain-groq>=0.1.0  # Groq for ultra-fast inference
langgraph>=0.0.40
anthropic
openai>=1.0.0
groq
google-generativeai  # Google Gemini models
deepagents  # Deep Agents library for hierarchical agents
tavily-python>=0.3.0  # Tavily web search API

# MIGRATION NOTE: Letta dependencies commented out - migrated to LangChain
# letta>=0.11.0        # Letta server and CLI commands (DEPRECATED)
# letta-client>=0.1.0  # Official Letta client SDK (DEPRECATED)
# sqlite-vec>=0.1.0    # SQLite vector extension for Letta (DEPRECATED)

# Tool-Based Architecture Dependencies
# helix-db>=0.1.0       # TODO: Phase 5 - Will implement custom knowledge storage
e2b-code-interpreter>=0.0.10  # Sandboxed code execution for analysis
firecrawl-py>=1.0.0     # Advanced web scraping for research

# Utilities
python-dotenv
requests
pydantic>=2.0.0
httpx
aiofiles
pyyaml>=6.0      # YAML configuration management